created components for logIn, logOut, and sign up
